Add files via upload
Initial files. I already created an html and css file that created the grid necessary to play the game. There are two boards, one for player 1 and another for player 2. The shogiMechanics.java has the initial functions for the necessary movement and actions of shogi as well as some extra functions like "threat. shogiMechanics.java might change in the future.